// src/types/UserDetails.ts
export type UserDetailsType = {
  age?: number | null;
  weight?: number | null;
  height?: number | null;
  chestCircumference?: number | null;
  waistCircumference?: number | null;
  hipCircumference?: number | null;
  armCircumference?: number | null;
  thighCircumference?: number | null;
  calfCircumference?: number | null;
  bmi?: number | null;
  bmr?: number | null;
  id?: number | string | null;
};